requirements.txt
fastapi
uvicorn[standard]
pydantic
radon
bandit
groq
python-dotenv
streamlit
requests